fix:better layout and control flow for vector tiles. chore:vector tile utils file.
