+++ src/TrackSegmentsLayer/index.test.js
@@ -21,14 +21,9 @@ jest.mock('../constants', () => ({
   MAP_ICON_SCALE: 2,
 }));
 
-jest.mock('../utils/tracks', () => ({
-  getVtRangeParam: (days) => {
-    const steps = [30, 45, 60, 90, 150, 210, 365, 500];
-    const step = steps.find((s) => days <= s);
-    return step !== undefined ? String(step) : 'all';
-  },
+jest.mock('../utils/vector-tiles', () => ({
+  ...jest.requireActual('../utils/vector-tiles'),
   buildVtTileUrl: (rangeParam) => `http://test-api.com/observations/segments/tiles/{z}/{x}/{y}.pbf?range=${rangeParam}`,
-  VECTOR_TILE_SOURCE: 'track-segments-source',
 }));
 
 jest.mock('react-redux', () => ({
